schbench: add per-cpu spin locking during the work
Real workloads have performance impacts to scheduling the process away
during critical sections.  We could spend a lot of time perfecting the
right set of cache misses to reflect those costs.

Or, we could make a per-cpu spinlock that gets held while we're doing
the matrix math.  This is a pretty blunt hammer but it accurately
reflects the impact of preemption on our workloads.

Signed-off-by: Chris Mason <clm@fb.com>