schbench: scale down worker threads to num_cpus by default
The default number of worker threads is just num_cpus, but the workers
are per message thread.  So if you pass -m 2, you'll end up with 2x the
number of workers, which is a surprise.

This commit changes the default to scale down the workers when -m is
increased.  You can still use -t to manually set the worker count.

Signed-off-by: Chris Mason <clm@fb.com>